added note creating with correct title, implemented note renaming

+++ src/utils/io.ts
@@ -3,12 +3,10 @@ import fs from 'fs-extra';
 import os from 'os';
 import chokidar from 'chokidar';
 import { app } from 'electron';
+
 // get application directory
 const appDir = path.resolve(app.getAppPath(), 'knowledge_base');
 
-/****************************/
-
-// get the list of files
 export const getFiles = () => {
 	fs.ensureDirSync(appDir);
 	const files = fs.readdirSync(appDir);
@@ -23,26 +21,22 @@ export const getFiles = () => {
 	});
 };
 
-/****************************/
-
-// add files
 export const addFile = (file: any) => {
 	// ensure `appDir` exists
 	fs.ensureDirSync(appDir);
-	const filePath = path.resolve(appDir, file.name);
-	if (!fs.existsSync(filePath)) {
-		fs.writeFileSync(filePath, file.data);
+	let filePath = path.resolve(appDir, file.name);
+	if (file.updateName) {
+		fs.renameSync(filePath, appDir + file.updateName);
+		filePath = path.resolve(appDir, file.updateName);
 	}
+	if (!fs.existsSync(filePath)) fs.writeFileSync(filePath, file.data);
 };
 
-// delete a file
 export const deleteFile = (filename: string) => {
 	const filePath = path.resolve(appDir, filename);
-	// remove file from the file system
 	if (fs.existsSync(filePath)) fs.removeSync(filePath);
 };
 
-// open a file
 export const readFile = (filename: string) => {
 	const filePath = path.resolve(appDir, filename);
 	if (fs.existsSync(filePath)) {
@@ -50,11 +44,7 @@ export const readFile = (filename: string) => {
 	}
 };
 
-/*-----*/
-
-// watch files from the application's storage directory
 export const watchFiles = (win: any) => {
-	//console.log(app.getAppPath());
 	chokidar.watch(appDir).on('unlink', (filepath) => {
 		win.webContents.send('app:delete-file', path.parse(filepath).base);
 	});
